+++ src/client/classes/base.class.js
@@ -1,5 +1,5 @@
 /*
- * pwix:forms/src/common/classes/base.class.js
+ * pwix:forms/src/client/classes/base.class.js
  *
  * A base class which does nothing but because '@vestergaard-company/js-mixin' only lets add an interface when extending a class...
  */