+++ src/common/js/functions.js
@@ -1,3 +1,0 @@
-/*
- * pwix:forms/src/common/js/functions.js
- */